requirements.txt
allure-pytest==2.14.0
allure-python-commons==2.14.0
annotated-types==0.7.0
anyio==4.4.0
attrs==25.1.0
certifi==2024.6.2
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
data==0.4
decorator==5.1.1
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
funcsigs==1.0.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
hupper==1.12.1
idna==3.7
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==2.2.2
orjson==3.10.5
outcome==1.3.0.post0
packaging==24.2
PasteDeploy==3.1.0
plaster==1.1.2
plaster-pastedeploy==1.0.1
player==0.6.1
pluggy==1.5.0
pycparser==2.22
pydantic==2.7.4
pydantic-extra-types==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.18.4
pyee==12.1.1
Pygments==2.18.0
pyramid==2.0.2
pyramid_jinja2==2.10.1
PySocks==1.7.1
pytest==8.3.5
pytest-base-url==2.1.0
python-dotenv==1.0.1
python-multipart==0.0.9
python-slugify==8.0.4
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
selenium==4.28.1
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.37.2
text-unidecode==1.3
translationstring==1.4
trio==0.29.0
trio-websocket==0.12.1
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.30.1
venusian==3.1.1
watchfiles==0.22.0
webdriver-manager==4.0.2
WebOb==1.8.9
websocket-client==1.8.0
websockets==12.0
wsproto==1.2.0
zope.deprecation==5.1
zope.interface==7.2